Test the buildpack in CI
Since previously the Travis run was a no-op, and we're about to add
support for a new stack (in the next PR), which would be good to be
able to test properly.

Uses an approach similar to that in:
https://github.com/heroku/heroku-buildpack-ci-postgresql

The Cedar-14 runs have been marked as `allow_failures`, since Chrome
currently fails to start due to:
`error while loading shared libraries: libgbm.so.1`

Fixing the buildpack on Cedar-14 is out of scope for this PR, plus the
stack is EOL and no-one has opened a support ticket for lack of Cedar-14
support yet.

Refs W-7502526.